--- sdk-python/copilotkit/demos/autotale_ai/demo.py
@@ -1,29 +1,0 @@
-"""Demo"""
-
-from fastapi import FastAPI
-import uvicorn
-from copilotkit.integrations.fastapi import add_fastapi_endpoint
-from copilotkit import CopilotKitSDK, LangGraphAgent
-from .agent import graph
-
-app = FastAPI()
-sdk = CopilotKitSDK(
-    agents=[
-        LangGraphAgent(
-            name="childrensBookAgent",
-            description="Write a children's book.",
-            graph=graph,
-        )
-    ],
-)
-
-add_fastapi_endpoint(app, sdk, "/copilotkit")
-
-def main():
-    """Run the uvicorn server."""
-    uvicorn.run(
-        "copilotkit.demos.autotale_ai.demo:app",
-        host="127.0.0.1",
-        port=8000,
-        reload=True
-    )
