--- sdk-python/copilotkit/demos/research_canvas/model.py
@@ -1,38 +1,0 @@
-"""
-This module provides a function to get a model based on the configuration.
-"""
-import os
-from typing import cast, Any
-from langchain_core.language_models.chat_models import BaseChatModel
-from copilotkit.demos.research_canvas.state import AgentState
-
-def get_model(state: AgentState) -> BaseChatModel:
-    """
-    Get a model based on the environment variable.
-    """
-
-    state_model = state.get("model")
-    model = os.getenv("MODEL", state_model)
-
-    print(f"Using model: {model}")
-
-    if model == "openai":
-        from langchain_openai import ChatOpenAI
-        return ChatOpenAI(temperature=0, model="gpt-4o-mini")
-    if model == "anthropic":
-        from langchain_anthropic import ChatAnthropic
-        return ChatAnthropic(
-            temperature=0,
-            model_name="claude-3-5-sonnet-20240620",
-            timeout=None,
-            stop=None
-        )
-    if model == "google_genai":
-        from langchain_google_genai import ChatGoogleGenerativeAI
-        return ChatGoogleGenerativeAI(
-            temperature=0,
-            model="gemini-1.5-pro",
-            api_key=cast(Any, os.getenv("GOOGLE_API_KEY")) or None
-        )
-
-    raise ValueError("Invalid model specified")
