--- sdk-python/copilotkit/demos/wait_user_input/demo.py
@@ -1,28 +1,0 @@
-"""Demo"""
-
-from dotenv import load_dotenv
-load_dotenv() # pylint: disable=wrong-import-position
-
-from fastapi import FastAPI
-import uvicorn
-from copilotkit.integrations.fastapi import add_fastapi_endpoint
-from copilotkit import CopilotKitSDK, LangGraphAgent
-from copilotkit.demos.wait_user_input.agent import graph
-
-
-app = FastAPI()
-sdk = CopilotKitSDK(
-    agents=[
-        LangGraphAgent(
-            name="weather_agent",
-            description="This agent deals with everything weather related",
-            graph=graph,
-        )
-    ],
-)
-
-add_fastapi_endpoint(app, sdk, "/copilotkit")
-
-def main():
-    """Run the uvicorn server."""
-    uvicorn.run("copilotkit.demos.wait_user_input.demo:app", host="127.0.0.1", port=8000, reload=True)
